Create seed script; add server; add db
In project-utils/seed.js, a seeding array is present, which the file
 then populates a database with if the database is empty. If the db
is full, then the final bundle name is logged.

The schema has been modified herein to reflect the space to be held
for the various values as well as to change the min column to "minim
um."

The database has been created and connected to sequelize. A simple
express server has also been created.