Respond to PR comments
In the db files, I've removed an unused DataTypes const as well as
 swapping the Sequelize & DataTypes consts in the model to a destr
uctured object.

In schema, I've fixed a nit to add a space following an =.

I've added a catch block to the server start in the server file.

Finally, this also adds a semicolon to fix a nit in the db spec.