Begin setup for components
This commit includes the additions of react, babel, webpack, and
the client folder in general w/ its own additions of an index.htm
l file as well as an index.jsx file to start working on my compon
ents.

I also added a few scripts to separate out my three testing conce
rns. If I've overinstalled dependencies in this setup, I'll delet
e prior to my PR for this code.