Update db & seeding for countdown
I've started getting this service ready to be the holder of the cou
ntdown clock, so I updated the schema, the sequelize model, & the d
atabase to account for that.

I've also changed the name of the database created in MySQL in orde
r not to accidentally drop anyone's database.